pyproject.toml
[project]
name = "fast-task"
version = "1.0"
description = ""
authors = [
    {name = "sabiqazhar", email = "sabiqandazhar@gmail.com"},
]
dependencies = [
    "fastapi>=0.110.0",
    "sqlalchemy>=2.0.29",
    "alembic>=1.13.1",
    "fastapi-sqlalchemy>=0.2.1",
    "pydantic>=2.6.4",
    "uvicorn>=0.29.0",
    "passlib>=1.7.4",
    "python-jose>=3.3.0",
    "pydantic-settings>=2.2.1",
    "psycopg2>=2.9.9",
    "psycopg2-binary>=2.9.9",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

